utils/text_extraction.py
import PyPDF2
from docx import Document

def extract_text(file):
    """Extract text from PDF or DOCX files"""
    if file.type == "application/pdf":
        reader = PyPDF2.PdfReader(file)
        return " ".join([page.extract_text() for page in reader.pages])
    elif file.type == "application/vnd.openxmlformats-officedocument.wordprocessingml.document":
        doc = Document(file)
        return " ".join([para.text for para in doc.paragraphs])
    else:
        return ""
