requirements.txt
streamlit
pandas
spacy
scikit-learn
python-docx
PyPDF2
matplotlib
seaborn



Install dependencies with:

pip install -r requirements.txt
python -m spacy download en_core_web_lg

